refactor(dry): reduce duplication between format and lint
Closes #24
