+++ package.json
@@ -40,15 +40,11 @@
     "@types/jest": "^29.5.1",
     "@types/validator": "^13.7.15",
     "jest": "^29.5.0",
-    "jest-extended": "^3.2.4",
     "ts-jest": "^29.1.0",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.4"
   },
   "jest": {
-    "setupFilesAfterEnv": [
-      "jest-extended/all"
-    ],
     "moduleFileExtensions": [
       "js",
       "json",
